[#76]/refactor useHandleBookmark로 삭제, 편집 함수 이동
